+++ CHANGELOG.rst
@@ -1,6 +1,18 @@
 Changelog
 =========
 
+dev
+---
+
+**Improvements**
+
+- Split terms in ``search`` command and match them separately
+- Add groups support when sharing a resource
+- Allow setting default recipients in the configuration file
+- Add support for tags
+- Add ``import_resources`` command
+- Make HTTP authentication optional in config wizard
+
 0.1.0 (2018-03-29)
 ------------------